fix pair all unit tests
